Crate saluki_app

High-level application primitives.

This crate provides common primitives necessary for bootstrapping an application prior to running, such as initializing logging, metrics, and memory management.

Modules

api

API server.

bootstrap

Bootstrap utilities.

config

Configuration API handler.

logging

Logging.

memory

Memory management.

metrics

Metrics.