saluki_app

Module logging

Logging.

Structs

LoggingAPIHandler

An API handler for updating log filtering directives at runtime.

LoggingHandlerState

State used for the logging API handler.

Functions

acquire_logging_api_handler

Acquires the logging API handler.

fatal_and_exit

Logs a message to standard error and exits the process with a non-zero exit code.

initialize_logging

Initializes the logging subsystem for tracing with the ability to dynamically update the log filtering directives at runtime.