saluki_app

Module metrics

Metrics.

Structs

MetricsAPIHandler

An API handler for updating metric filtering directives at runtime.

Functions

acquire_metrics_api_handler

Acquires the metrics API handler.

collect_runtime_metrics

Collects Tokio runtime metrics on the current Tokio runtime.

emit_startup_metrics

Emits the startup metrics for the application.